+++ Quiz/Backend/src/main/java/com/example/backend/quiz/QuizMapper.java
@@ -30,7 +30,7 @@ public class QuizMapper {
             AnswerEntity[] answers = gson.fromJson(answersString, AnswerEntity[].class);
             int maxtime = queryResults.getInt(6);
             int value = queryResults.getInt(7);
-            List<Integer> order = gson.fromJson(queryResults.getString(8), List.class);
+            List<Double> order = gson.fromJson(queryResults.getString(8), List.class);
             int questionId = queryResults.getInt(9);
 
             List<QuizEntity> firstQuizCollected = quizEntities.stream().filter(q -> q.getId() == quizId).toList();
